+++ tsup.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
   clean: true,
   external: ['react', 'react-dom'],
   treeshake: true,
-  sourcemap: true,
+  sourcemap: false,
   loader: {
     '.css': 'local-css',
   },